requirements.txt
airspeed==0.5.5.dev20160812
asn1crypto==0.23.0
aws-xray-sdk==0.93
awscli==1.11.185
awscli-local==0.4
boto==2.46.1
boto3==1.4.7
botocore==1.7.43
cached-property==1.3.1
cachetools==0.8.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
colorama==0.3.7
configparser==3.5.0
cookies==2.2.1
coverage==4.4.2
cryptography==2.1.3
decorator==4.1.2
dnslib==0.9.7
docker==2.6.1
docker-compose==1.13.0
docker-pycreds==0.2.1
docker-test-tools==1.0.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
ecdsa==0.13
elasticsearch==5.3.0
envs==1.2.3
fancycompleter==0.8
flake8==3.5.0
flake8-quotes==0.12.0
Flask==0.10.1
Flask-Cors==3.0.3
flask-swagger==0.2.12
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonpath-rw==1.4.0
jsonpickle==0.9.5
jsonschema==2.6.0
localstack==0.8.2
localstack-client==0.4
localstack-ext==0.7.0.7
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
moto-ext==1.1.24
nose==1.3.7
nose2==0.7.0
pbr==1.8.0
pdbpp==0.9.2
pep8==1.7.1
ply==3.10
psutil==5.2.0
pyaes==1.6.0
pyaml==17.10.0
pyasn1==0.3.7
pycodestyle==2.3.1
pycparser==2.18
pycryptodome==3.3.1
pyflakes==1.6.0
pyforge==1.3.0
Pygments==2.2.0
pyminifier==2.1
pyOpenSSL==17.0.0
pyspark==2.1.1+hadoop2.7
python-coveralls==2.7.0
python-dateutil==2.6.1
python-dotenv==0.7.1
python-jose-ext==1.3.2.2
pytz==2017.3
PyYAML==3.12
requests==2.13.0
requests-aws4auth==0.9
requests-unixsocket==0.1.5
rsa==3.4.2
s3transfer==0.1.11
sentinels==1.0.0
sh==1.12.14
six==1.11.0
spark-parser==1.8.4
srp-ext==1.0.7.1
subprocess32-ext==3.2.8.2
texttable==0.8.8
urllib3==1.22
waiting==1.3.0
warrant-ext==0.2.0.3
websocket-client==0.44.0
Werkzeug==0.12.2
wmctrl==0.3
wrapt==1.10.11
xdis==3.6.1
xmltodict==0.10.2
